requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
bcrypt==4.0.0
click==8.3.2
fastapi==0.136.0
greenlet==3.4.0
h11==0.16.0
idna==3.11
Mako==1.3.11
MarkupSafe==3.0.3
passlib==1.7.4
psycopg==3.3.3
psycopg-binary==3.3.3
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
PyJWT==2.12.1
python-dotenv==1.2.2
scalar_fastapi==1.8.2
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
